CTCP-1466: Update validation code error passing to support sending back parsing errors in Json
This brings the error handling code in line with other projects. Note that because a validation error is considered an error in the services, but not in the controller, we have to shift some of the errors from the left to the right to account for what we want to output.